refactor(button): change naming convention
