+++ stories/Button.stories.mdx
@@ -2,12 +2,12 @@ import { Canvas, Meta, Story, ArgsTable } from "@storybook/addon-docs";
 import { html } from "lit";
 import { withWebComponentsKnobs } from "storybook-addon-web-components-knobs";
 import { withKnobs, text, number } from "@storybook/addon-knobs";
-import { ButtonElement } from "../lib/Button";
+import "../lib/Button";
 
 <Meta
   title="Components/Button"
   argTypes={{
-    classes: { control: "text" },
+    buttonclasses: { control: "text" },
     label: { 
       control: "text",
       defaultValue: 'Hello World',
@@ -78,9 +78,9 @@ import { ButtonElement } from "../lib/Button";
 
 export const Template = (args) =>
   html`
-    <button-element 
+    <sgds-button 
       .variant=${args.variant} 
-      .classes=${args.classes} 
+      .buttonclasses=${args.buttonclasses} 
       .size=${args.size}
       .active=${args.active}
       .disabled=${args.disabled}
@@ -89,10 +89,10 @@ export const Template = (args) =>
       .download=${args.download}
     >
       <span slot="label">${args.label}</span>
-    </button-element>
+    </sgds-button>
   `;
 
-# ButtonElement
+# SgdsButton
 
 Buttons communicate actions that users can take throughout your portal which can be used in places such as modals, forms and cards
 
@@ -105,7 +105,7 @@ Buttons communicate actions that users can take throughout your portal which can
 ## API
 
 ```jsx
-import { ButtonElement } from "@govtechsg/sgds-web-component";
+import { SgdsButton } from "@govtechsg/sgds-web-component";
 ```
 
 <ArgsTable story="Basic" />
